secret-scan/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "secret-scan"
version = "0.1.0"
description = "A tool for scanning secrets in code repositories"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "secret-scan team" }
]
keywords = ["security", "secret-scanning", "devsecops"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "click>=8.0",
  "pyyaml>=6.0",
  "jinja2>=3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
]

[project.scripts]
secret-scan = "secret_scan.__main__:cli"

[tool.hatch.build.targets.wheel]
packages = ["secret_scan"]
include = [
  "secret_scan/builtin_rules.yml",
  "secret_scan/templates/",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.uv.workspace]
members = [
    ".",
    ".",
]

[tool.uv.sources]
secret-scan = { workspace = true }

[dependency-groups]
dev = [
    "secret-scan",
]
